+++ Editor/XRGeneralSettingsPerBuildTarget.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using System.Threading;
 using UnityEngine;
 using UnityEngine.XR.Management;
+using UnityEditor.PackageManager;
 
 using UnityEditor.XR.Management.Metadata;
 
@@ -59,12 +61,28 @@ namespace UnityEditor.XR.Management
                 AssetDatabase.CreateAsset(generalSettings, assetPath);
                 AssetDatabase.SaveAssets();
             }
+
             EditorBuildSettings.AddConfigObject(XRGeneralSettings.k_SettingsKey, generalSettings, true);
             return generalSettings;
         }
 
         internal static XRGeneralSettingsPerBuildTarget GetOrCreate()
-            => TryFindSettingsAsset(out var generalSettings) ? generalSettings : CreateAssetSynchronized();
+        {
+            TryFindSettingsAsset(out var generalSettings);
+            generalSettings = generalSettings ?? CreateAssetSynchronized();
+
+            var buildTargetSettings = generalSettings.SettingsForBuildTarget(
+                EditorUserBuildSettings.selectedBuildTargetGroup
+            );
+            var activeLoaderCount = (buildTargetSettings?.AssignedSettings?.activeLoaders?.Count ?? 0);
+            if (IsInURPGraphicsTest() && activeLoaderCount == 0)
+            {
+                Debug.Log("In URP graphics test and no loaders selected. Disabling 'Initialize XR On Startup'");
+                buildTargetSettings.InitManagerOnStart = false;
+            }
+
+            return generalSettings;
+        }
 
         // Simple class to give us updates when the asset database changes.
         class AssetCallbacks : AssetPostprocessor
@@ -134,6 +152,43 @@ namespace UnityEditor.XR.Management
             }
 
             return false;
+        }
+
+        // Temporary workaround for a graphics issue
+        static bool? s_IsInURPGraphicsTest = null;
+        internal static bool IsInURPGraphicsTest()
+        {
+            if (!s_IsInURPGraphicsTest.HasValue)
+            {
+                s_IsInURPGraphicsTest = false;
+
+                const string kGfxFramework = "com.unity.testing.urp";
+
+                var request = Client.List(offlineMode: true);
+                while (request.IsCompleted == false)
+                {
+                    Thread.Yield();
+                }
+
+                if (request.Status != StatusCode.Success)
+                {
+                    throw new InvalidOperationException(
+                        $"Unexpected package query failure searching for {kGfxFramework}"
+                    );
+                }
+
+                // if we find the com.unity.testing.urp package then assume we're in the URP
+                // graphics test.
+                foreach (var result in request.Result)
+                {
+                    if (result.packageId.StartsWith(kGfxFramework))
+                    {
+                        s_IsInURPGraphicsTest = true;
+                        break;
+                    }
+                }
+            }
+            return s_IsInURPGraphicsTest.Value;
         }
 #endif