+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.3.3] - 2023-02-14
+
+### Fixed
+- Settings and Loader names with numbers will now keep the numbers for the generated asset names.
+
+## [4.3.2] - 2023-02-06
+
+### Fixed
+- Incompatibility with `com.unity.testframework.graphics`
+
 ## [4.3.1] - 2022-10-14
 
 ### Fixed